Remove null service protections from pipes
This has been done because the effect of unregistered services is far more obvious than the previous behaviour. Previously, the pipes would generally continue, mimicking a successful response. This means that invalid setups previously lead to successful responses, instead of failures.

Missing registrations can easily be found by writing integration tests on the use case, which is preferable to writing unit tests on individual components of the use case. It's preferable because there's a lot of behaviour implemented in the pipes, and the interaction of pipes and their services isn't tested through unit tests alone.

#126