Migrate PipeHandle initialisation to InvokeAsync
This is because the ServiceFactory is registered as a scoped service and the Pipeline is registered as a singleton service. Injecting a scoped service into a singleton service is bad practice and will fail when registering in certain DI containers. Benchmarks for a simplified situation indicate that the if check in the InvokeAsync method will only add a nanosecond per invocation.

#147